Move files to liveshare folder
As per #641, liveshare files are moved to a /liveshare/ folder, and the "r" prefix has been removed from their name
