+++ package.json
@@ -111,7 +111,7 @@
       },
       {
         "view": "rLiveShare",
-        "contents": "R LiveShare not active."
+        "contents": "R Live Share not active."
       }
     ],
     "languages": [
